requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
python-dotenv==1.0.1
cachetools==5.3.3
pytz==2024.1
notion-client
pytest-asyncio
databases[sqlite,postgresql]
asyncpg
google-auth
requests
